# docs/swagger.yaml
definitions:
...
  server.updateMeetingActionItemPayload:
    properties:
      remarks:
        type: string
      status:
        type: string
    type: object
...
host: localhost:4000
info:
  contact: {}
  description: REST API aplikasi EBITDA Max APN (refactor Go + Next.js) — scope manager
    KDKMP & superadmin.
  title: EBITDA Max APN API
  version: "1.0"
paths:
...
  /api/v1/meeting-minutes/action-items:
    get:
      parameters:
      - description: Cari subjek, aksi, PIC, atau meeting
        in: query
        name: search
        type: string
      - description: open|in_progress|completed|cancelled
        in: query
        name: status
        type: string
      - description: Hanya item terlambat
        in: query
        name: overdue
        type: boolean
      - description: Halaman
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Daftar Action Items milik Manager KDKMP
      tags:
      - Meeting Minutes
  /api/v1/meeting-minutes/action-items/{itemID}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID Action Item
        in: path
        name: itemID
        required: true
        type: integer
      - description: Status dan catatan
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/server.updateMeetingActionItemPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Perbarui status Action Item milik Manager KDKMP
      tags:
      - Meeting Minutes
...
schemes:
- http
securityDefinitions:
  CookieAuth:
    in: cookie
    name: ebitda_session
    type: apiKey
swagger: "2.0"